add: `utility` directory, constant.dart and logger.dart move to here
