+++ github_search/lib/presentation/github_search_screen/widget/github_search_text_field_widget.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:github_search/logger.dart';
 
 class GitHubSearchTextFieldWidget extends StatelessWidget {
   const GitHubSearchTextFieldWidget({
