+++ github_search/lib/repository/network/request/github_search_request.dart
@@ -1,5 +1,5 @@
 
-import 'package:github_search/constant.dart';
+import 'package:github_search/utility/constant.dart';
 import 'package:github_search/repository/network/http_method.dart';
 import 'package:github_search/repository/network/request/request_interface.dart';
